+++ examples/embeddable_examples/public/app/register_embeddable.tsx
@@ -81,13 +81,14 @@ export const RegisterEmbeddable = () => {
       <EuiSpacer size="l" />
 
       <EuiText>
-        <h2>Saved object embeddables</h2>
+        <h2>
+          Show saved object type in <em>Add from library</em> menu
+        </h2>
         <p>
           Embeddable factories, such as Lens, Maps, Links, that can reference saved objects should
-          register their saved object types using{' '}
-          <strong>registerReactEmbeddableSavedObject</strong>. The <em>Add from library</em> flyout
-          on Dashboards uses this registry to list saved objects. The example function below could
-          be called from the public start contract for a plugin.
+          register their saved object types using <strong>registerAddFromLibraryType</strong>. The{' '}
+          <em>Add from library</em> flyout on Dashboards uses this registry to list saved objects.
+          The example function below could be called from the public start contract for a plugin.
         </p>
       </EuiText>
       <EuiSpacer size="s" />
