+++ src/plugins/embeddable/public/index.ts
@@ -10,7 +10,8 @@
 import { PluginInitializerContext } from '@kbn/core/public';
 import { EmbeddablePublicPlugin } from './plugin';
 
-export { openAddPanelFlyout } from './add_panel_flyout/open_add_panel_flyout';
+export { useAddFromLibraryTypes } from './add_from_library/registry';
+export { openAddFromLibraryFlyout } from './add_from_library/open_add_from_library_flyout';
 export { EmbeddablePanel } from './embeddable_panel';
 export {
   cellValueTrigger,
@@ -80,7 +81,6 @@ export type {
   PanelState,
   PropertySpec,
   RangeSelectContext,
-  ReactEmbeddableSavedObject,
   ReferenceOrValueEmbeddable,
   SavedObjectEmbeddableInput,
   SelfStyledEmbeddable,
