+++ src/plugins/embeddable/public/lib/index.ts
@@ -17,4 +17,3 @@ export * from './reference_or_value_embeddable';
 export * from './self_styled_embeddable';
 export * from './filterable_embeddable';
 export * from './factory_migrations/run_factory_migrations';
-export * from './embeddable_saved_object_registry';
