+++ src/plugins/embeddable/public/mocks.tsx
@@ -33,14 +33,13 @@ import {
   SelfStyledEmbeddable,
 } from '.';
 import { setKibanaServices } from './kibana_services';
-import { registerReactEmbeddableSavedObject } from './lib';
 import { SelfStyledOptions } from './lib/self_styled_embeddable/types';
 import { EmbeddablePublicPlugin } from './plugin';
 import {
   reactEmbeddableRegistryHasKey,
   registerReactEmbeddableFactory,
 } from './react_embeddable_system';
-import { registerSavedObjectToPanelMethod } from './registry/saved_object_to_panel_methods';
+import { registerAddFromLibraryType } from './add_from_library/registry';
 
 export { mockAttributeService } from './lib/attribute_service/attribute_service.mock';
 export type Setup = jest.Mocked<EmbeddableSetup>;
@@ -100,12 +99,7 @@ export function mockFilterableEmbeddable<OriginalEmbeddableType>(
 
 const createSetupContract = (): Setup => {
   const setupContract: Setup = {
-    registerSavedObjectToPanelMethod: jest
-      .fn()
-      .mockImplementation(registerSavedObjectToPanelMethod),
-    registerReactEmbeddableSavedObject: jest
-      .fn()
-      .mockImplementation(registerReactEmbeddableSavedObject),
+    registerAddFromLibraryType: jest.fn().mockImplementation(registerAddFromLibraryType),
     registerReactEmbeddableFactory: jest.fn().mockImplementation(registerReactEmbeddableFactory),
     registerEmbeddableFactory: jest.fn(),
     registerEnhancement: jest.fn(),
@@ -117,7 +111,6 @@ const createSetupContract = (): Setup => {
 const createStartContract = (): Start => {
   const startContract: Start = {
     reactEmbeddableRegistryHasKey: jest.fn().mockImplementation(reactEmbeddableRegistryHasKey),
-    getReactEmbeddableSavedObjects: jest.fn(),
     getEmbeddableFactories: jest.fn(),
     getEmbeddableFactory: jest.fn(),
     telemetry: jest.fn(),
