+++ src/plugins/embeddable/public/plugin.tsx
@@ -25,7 +25,6 @@ import { migrateToLatest, PersistableStateService } from '@kbn/kibana-utils-plug
 import { SavedObjectsManagementPluginStart } from '@kbn/saved-objects-management-plugin/public';
 import type { ContentManagementPublicStart } from '@kbn/content-management-plugin/public';
 import type { SavedObjectTaggingOssPluginStart } from '@kbn/saved-objects-tagging-oss-plugin/public';
-import { FinderAttributes } from '@kbn/saved-objects-finder-plugin/common';
 import {
   EmbeddableFactoryRegistry,
   EmbeddableFactoryProvider,
@@ -41,9 +40,6 @@ import {
   defaultEmbeddableFactoryProvider,
   IEmbeddable,
   SavedObjectEmbeddableInput,
-  registerReactEmbeddableSavedObject,
-  ReactEmbeddableSavedObject,
-  getReactEmbeddableSavedObjects,
 } from './lib';
 import { EmbeddableFactoryDefinition } from './lib/embeddables/embeddable_factory_definition';
 import { EmbeddableStateTransfer } from './lib/state_transfer';
@@ -62,7 +58,7 @@ import {
   reactEmbeddableRegistryHasKey,
   registerReactEmbeddableFactory,
 } from './react_embeddable_system';
-import { registerSavedObjectToPanelMethod } from './registry/saved_object_to_panel_methods';
+import { registerAddFromLibraryType } from './add_from_library/registry';
 
 export interface EmbeddableSetupDependencies {
   uiActions: UiActionsSetup;
@@ -79,17 +75,16 @@ export interface EmbeddableStartDependencies {
 
 export interface EmbeddableSetup {
   /**
-   * Register an embeddable API saved object with the Add from library flyout.
+   * Register a saved object type with the "Add from library" flyout.
    *
    * @example
-   *  registerReactEmbeddableSavedObject({
+   *  registerAddFromLibraryType({
    *    onAdd: (container, savedObject) => {
    *      container.addNewPanel({
    *        panelType: CONTENT_ID,
    *        initialState: savedObject.attributes,
    *      });
    *    },
-   *    embeddableType: CONTENT_ID,
    *    savedObjectType: MAP_SAVED_OBJECT_TYPE,
    *    savedObjectName: i18n.translate('xpack.maps.mapSavedObjectLabel', {
    *      defaultMessage: 'Map',
@@ -97,12 +92,7 @@ export interface EmbeddableSetup {
    *    getIconForSavedObject: () => APP_ICON,
    *  });
    */
-  registerReactEmbeddableSavedObject: typeof registerReactEmbeddableSavedObject;
-
-  /**
-   * @deprecated React embeddables should register their saved objects with {@link registerReactEmbeddableSavedObject}.
-   */
-  registerSavedObjectToPanelMethod: typeof registerSavedObjectToPanelMethod;
+  registerAddFromLibraryType: typeof registerAddFromLibraryType;
 
   /**
    * Registers an async {@link ReactEmbeddableFactory} getter.
@@ -135,14 +125,6 @@ export interface EmbeddableStart extends PersistableStateService<EmbeddableState
    * Checks if a {@link ReactEmbeddableFactory} has been registered using {@link registerReactEmbeddableFactory}
    */
   reactEmbeddableRegistryHasKey: (type: string) => boolean;
-
-  /**
-   *
-   * @returns An iterator over all {@link ReactEmbeddableSavedObject}s that have been registered using {@link registerReactEmbeddableSavedObject}.
-   */
-  getReactEmbeddableSavedObjects: <
-    TSavedObjectAttributes extends FinderAttributes
-  >() => IterableIterator<[string, ReactEmbeddableSavedObject<TSavedObjectAttributes>]>;
 
   /**
    * @deprecated use {@link registerReactEmbeddableFactory} instead.
@@ -192,8 +174,7 @@ export class EmbeddablePublicPlugin implements Plugin<EmbeddableSetup, Embeddabl
 
     return {
       registerReactEmbeddableFactory,
-      registerSavedObjectToPanelMethod,
-      registerReactEmbeddableSavedObject,
+      registerAddFromLibraryType,
 
       registerEmbeddableFactory: this.registerEmbeddableFactory,
       registerEnhancement: this.registerEnhancement,
@@ -244,7 +225,6 @@ export class EmbeddablePublicPlugin implements Plugin<EmbeddableSetup, Embeddabl
 
     const embeddableStart: EmbeddableStart = {
       reactEmbeddableRegistryHasKey,
-      getReactEmbeddableSavedObjects,
 
       getEmbeddableFactory: this.getEmbeddableFactory,
       getEmbeddableFactories: this.getEmbeddableFactories,
