+++ src/plugins/visualizations/public/plugin.ts
@@ -411,14 +411,13 @@ export class VisualizationsPlugin
       const getVisualizeEmbeddableFactory = await getVisualizeEmbeddableFactoryLazy();
       return getVisualizeEmbeddableFactory({ embeddableStart, embeddableEnhancedStart });
     });
-    embeddable.registerReactEmbeddableSavedObject<VisualizationSavedObjectAttributes>({
+    embeddable.registerAddFromLibraryType<VisualizationSavedObjectAttributes>({
       onAdd: (container, savedObject) => {
         container.addNewPanel<VisualizeSerializedState>({
           panelType: VISUALIZE_EMBEDDABLE_TYPE,
           initialState: { savedObjectId: savedObject.id },
         });
       },
-      embeddableType: VISUALIZE_EMBEDDABLE_TYPE,
       savedObjectType: VISUALIZE_EMBEDDABLE_TYPE,
       savedObjectName: i18n.translate('visualizations.visualizeSavedObjectName', {
         defaultMessage: 'Visualization',
