+++ x-pack/plugins/canvas/public/components/embeddable_flyout/flyout.component.tsx
@@ -9,14 +9,9 @@ import { EuiFlyout, EuiFlyoutBody, EuiFlyoutHeader, EuiTitle } from '@elastic/eu
 import { i18n } from '@kbn/i18n';
 import React, { FC, useCallback, useMemo } from 'react';
 
-import { EmbeddableFactory, ReactEmbeddableSavedObject } from '@kbn/embeddable-plugin/public';
-import { FinderAttributes } from '@kbn/saved-objects-finder-plugin/common';
-import { SavedObjectFinder, SavedObjectMetaData } from '@kbn/saved-objects-finder-plugin/public';
-import {
-  contentManagementService,
-  coreServices,
-  embeddableService,
-} from '../../services/kibana_services';
+import { useAddFromLibraryTypes } from '@kbn/embeddable-plugin/public';
+import { SavedObjectFinder } from '@kbn/saved-objects-finder-plugin/public';
+import { contentManagementService, coreServices } from '../../services/kibana_services';
 
 const strings = {
   getNoItemsText: () =>
@@ -28,13 +23,6 @@ const strings = {
       defaultMessage: 'Add from library',
     }),
 };
-
-interface LegacyFactoryMap {
-  [key: string]: EmbeddableFactory;
-}
-interface FactoryMap<TSavedObjectAttributes extends FinderAttributes = FinderAttributes> {
-  [key: string]: ReactEmbeddableSavedObject<TSavedObjectAttributes> & { type: string };
-}
 
 export interface Props {
   onClose: () => void;
@@ -49,81 +37,18 @@ export const AddEmbeddableFlyout: FC<Props> = ({
   onClose,
   isByValueEnabled,
 }) => {
-  const legacyFactoriesBySavedObjectType: LegacyFactoryMap = useMemo(() => {
-    return [...embeddableService.getEmbeddableFactories()]
-      .filter(
-        (embeddableFactory) =>
-          Boolean(embeddableFactory.savedObjectMetaData?.type) && !embeddableFactory.isContainerType
-      )
-      .reduce((acc, factory) => {
-        acc[factory.savedObjectMetaData!.type] = factory;
-        return acc;
-      }, {} as LegacyFactoryMap);
-  }, []);
+  const libraryTypes = useAddFromLibraryTypes();
 
-  const factoriesBySavedObjectType: FactoryMap = useMemo(() => {
-    return [...embeddableService.getReactEmbeddableSavedObjects()]
-      .filter(([type, embeddableFactory]) => {
-        return Boolean(embeddableFactory.savedObjectMetaData?.type);
-      })
-      .reduce((acc, [type, factory]) => {
-        acc[factory.savedObjectMetaData!.type] = {
-          ...factory,
-          type,
-        };
-        return acc;
-      }, {} as FactoryMap);
-  }, []);
-
-  const metaData = useMemo(
-    () =>
-      [
-        ...Object.values(factoriesBySavedObjectType),
-        ...Object.values(legacyFactoriesBySavedObjectType),
-      ]
-        .filter((factory) =>
-          Boolean(
-            factory.type !== 'links' && // Links panels only exist on Dashboards
-              (isByValueEnabled || availableEmbeddables.includes(factory.type))
-          )
-        )
-        .map((factory) => factory.savedObjectMetaData)
-        .filter<SavedObjectMetaData<{}>>(function (
-          maybeSavedObjectMetaData
-        ): maybeSavedObjectMetaData is SavedObjectMetaData<{}> {
-          return maybeSavedObjectMetaData !== undefined;
-        })
-        .sort((a, b) => a.type.localeCompare(b.type)),
-    [
-      availableEmbeddables,
-      factoriesBySavedObjectType,
-      isByValueEnabled,
-      legacyFactoriesBySavedObjectType,
-    ]
-  );
+  const canvasOnlyLibraryTypes = useMemo(() => {
+    // Links panels are not supported in Canvas
+    return libraryTypes.filter(({ type }) => type !== 'links');
+  }, [libraryTypes]);
 
   const onAddPanel = useCallback(
     (id: string, savedObjectType: string) => {
-      if (factoriesBySavedObjectType[savedObjectType]) {
-        const factory = factoriesBySavedObjectType[savedObjectType];
-        const { type } = factory;
-        onSelect(id, type, isByValueEnabled);
-        return;
-      }
-      const embeddableFactories = embeddableService.getEmbeddableFactories();
-      // Find the embeddable type from the saved object type
-      const found = Array.from(embeddableFactories).find((embeddableFactory) => {
-        return Boolean(
-          embeddableFactory.savedObjectMetaData &&
-            embeddableFactory.savedObjectMetaData.type === savedObjectType
-        );
-      });
-
-      const foundEmbeddableType = found ? found.type : 'unknown';
-
-      onSelect(id, foundEmbeddableType, isByValueEnabled);
+      onSelect(id, savedObjectType, isByValueEnabled);
     },
-    [isByValueEnabled, onSelect, factoriesBySavedObjectType]
+    [isByValueEnabled, onSelect]
   );
 
   return (
@@ -137,7 +62,7 @@ export const AddEmbeddableFlyout: FC<Props> = ({
         <SavedObjectFinder
           id="canvasEmbeddableFlyout"
           onChoose={onAddPanel}
-          savedObjectMetaData={metaData}
+          savedObjectMetaData={canvasOnlyLibraryTypes}
           showFilter={true}
           noItemsMessage={strings.getNoItemsText()}
           services={{
